+++ apps/desktop/src/components/right-panel/hooks/useChatLogic.ts
@@ -4,7 +4,6 @@ import { message } from "@tauri-apps/plugin-dialog";
 import { useCallback, useEffect, useRef, useState } from "react";
 
 import type { SelectionData } from "@/contexts/right-panel";
-import { globalEditorRef } from "../../../shared/editor-ref";
 
 import { useLicense } from "@/hooks/use-license";
 import { commands as analyticsCommands } from "@hypr/plugin-analytics";
@@ -29,6 +28,8 @@ import type { ActiveEntityInfo, Message } from "../types/chat-types";
 import { prepareMessageHistory } from "../utils/chat-utils";
 import { parseMarkdownBlocks } from "../utils/markdown-parser";
 import { buildVercelToolsFromMcp } from "../utils/mcp-http-wrapper";
+import { createSearchSessionTool } from "../utils/tools/search_session_multi_keywords";
+import { createEditEnhancedNoteTool } from "../utils/tools/edit_enhanced_note";
 
 
 interface UseChatLogicProps {
@@ -288,293 +289,12 @@ export function useChatLogic({
         }
       }
 
-      const searchTool = tool({
-        description:
-          "Search for sessions (meeting notes) with multiple keywords. The keywords should be the most important things that the user is talking about. This could be either topics, people, or company names.",
-        inputSchema: z.object({
-          keywords: z.array(z.string()).min(3).max(5).describe(
-            "List of 3-5 keywords to search for, each keyword should be concise",
-          ),
-        }),
-        execute: async ({ keywords }) => {
-          const searchPromises = keywords.map(keyword =>
-            dbCommands.listSessions({
-              type: "search",
-              query: keyword,
-              user_id: userId || "",
-              limit: 3,
-            })
-          );
-
-          const searchResults = await Promise.all(searchPromises);
-
-          const combinedResults = new Map();
-
-          searchResults.forEach((sessions, index) => {
-            const keyword = keywords[index];
-            sessions.forEach(session => {
-              if (combinedResults.has(session.id)) {
-                combinedResults.get(session.id).matchedKeywords.push(keyword);
-              } else {
-                combinedResults.set(session.id, {
-                  ...session,
-                  matchedKeywords: [keyword],
-                });
-              }
-            });
-          });
-
-          const finalResults = Array.from(combinedResults.values())
-            .sort((a, b) => b.matchedKeywords.length - a.matchedKeywords.length);
-
-          return {
-            results: finalResults,
-            summary: {
-              totalSessions: finalResults.length,
-              keywordsSearched: keywords,
-              sessionsByKeywordCount: finalResults.reduce((acc, session) => {
-                const count = session.matchedKeywords.length;
-                acc[count] = (acc[count] || 0) + 1;
-                return acc;
-              }, {} as Record<number, number>),
-            },
-          };
-        },
-      });
-
-      const editEnhancedNoteTool = tool({
-        description: "Edit a specific part of the enhanced note by replacing HTML content at given ProseMirror positions with new HTML content. Use this when the user asks to modify, change, or replace specific selected content. The selected content is provided as HTML, and you should respond with HTML that maintains proper formatting.",
-        inputSchema: z.object({
-          startOffset: z.number().describe("The ProseMirror start position of the content to replace"),
-          endOffset: z.number().describe("The ProseMirror end position of the content to replace"), 
-          newHtml: z.string().describe("The new HTML content to replace the selected content with. Maintain proper HTML structure and formatting."),
-        }),
-        execute: async ({ startOffset, endOffset, newHtml }) => {
-          if (!sessionId) {
-            return { success: false, error: "No session ID available" };
-          }
-
-          const sessionStore = sessions[sessionId];
-          if (!sessionStore) {
-            return { success: false, error: "Session not found" };
-          }
-
-          try {
-            const editor = globalEditorRef.current;
-            
-            if (!editor) {
-              return { success: false, error: "Editor not available" };
-            }
-
-            // Trim and clean the HTML to prevent empty elements
-            const cleanedHtml = newHtml.trim().replace(/>\s+</g, '><');
-            
-            // Store initial doc size for accurate position calculation
-            const initialDocSize = editor.state.doc.content.size;
-            
-            // Delete old content and insert new content with proper parse options
-            editor.chain()
-              .focus()
-              .setTextSelection({ from: startOffset, to: endOffset })
-              .deleteSelection()
-              .insertContent(cleanedHtml, {
-                parseOptions: {
-                  preserveWhitespace: false
-                }
-              })
-              .run();
-            
-            // Calculate actual inserted content size
-            const finalDocSize = editor.state.doc.content.size;
-            const insertedSize = finalDocSize - initialDocSize + (endOffset - startOffset);
-            const highlightEnd = startOffset + insertedSize;
-            
-            // Apply AI highlight to the actually inserted content with metadata
-            editor.chain()
-              .setTextSelection({ from: startOffset, to: highlightEnd })
-              .setAIHighlight({
-                timestamp: Date.now().toString(),
-                sessionId: sessionId || undefined
-              })
-              .run();
-
-            // Create simple floating accept/undo controls
-            const createControls = () => {
-              const highlighted = document.querySelector('[data-ai-highlight="true"]');
-              if (!highlighted) return;
-
-              // Remove any existing controls
-              document.querySelector('.ai-edit-controls')?.remove();
-
-              const rect = highlighted.getBoundingClientRect();
-              const controls = document.createElement('div');
-              controls.className = 'ai-edit-controls';
-              controls.style.cssText = `
-                position: fixed;
-                top: ${rect.top - 36}px;
-                left: ${rect.left}px;
-                z-index: 9999;
-                background: white;
-                border: 1px solid #e5e7eb;
-                border-radius: 6px;
-                padding: 2px;
-                display: flex;
-                gap: 2px;
-                box-shadow: 0 2px 6px rgba(0,0,0,0.1);
-              `;
-
-              // Undo button
-              const undoBtn = document.createElement('button');
-              undoBtn.innerHTML = `
-                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
-                  <path d="M3 7v6h6"/><path d="M21 17a9 9 0 00-9-9 9 9 0 00-6 2.3L3 13"/>
-                </svg>
-                <span style="margin-left: 4px; font-size: 11px;">Undo</span>
-              `;
-              undoBtn.style.cssText = `
-                display: flex;
-                align-items: center;
-                padding: 4px 8px;
-                background: white;
-                border: none;
-                border-radius: 4px;
-                cursor: pointer;
-                font-size: 12px;
-                color: #374151;
-              `;
-              undoBtn.onmouseover = () => undoBtn.style.background = '#f3f4f6';
-              undoBtn.onmouseout = () => undoBtn.style.background = 'white';
-              undoBtn.onclick = () => {
-                // Focus the actual editor DOM element first
-                const editorElement = document.querySelector('.tiptap-normal, .ProseMirror');
-                if (editorElement instanceof HTMLElement) {
-                  editorElement.focus();
-                  
-                  // Wait for focus to take effect, then undo
-                  setTimeout(() => {
-                    const currentEditor = globalEditorRef.current;
-                    if (currentEditor) {
-                      // Check if undo is available
-                      if (currentEditor.can().undo()) {
-                        currentEditor.commands.undo();
-                      } else {
-                        console.warn('Undo not available');
-                      }
-                    } else {
-                      console.warn('No editor reference available');
-                    }
-                    (controls as any).cleanup?.() || controls.remove();
-                  }, 50);
-                } else {
-                  console.warn('Could not find editor element to focus');
-                  (controls as any).cleanup?.() || controls.remove();
-                }
-              };
-
-              // Accept button
-              const acceptBtn = document.createElement('button');
-              acceptBtn.innerHTML = `
-                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="white" stroke-width="2">
-                  <polyline points="20 6 9 17 4 12"/>
-                </svg>
-                <span style="margin-left: 4px; font-size: 11px;">Accept</span>
-              `;
-              acceptBtn.style.cssText = `
-                display: flex;
-                align-items: center;
-                padding: 4px 8px;
-                background: #3b82f6;
-                border: none;
-                border-radius: 4px;
-                cursor: pointer;
-                font-size: 12px;
-                color: white;
-              `;
-              acceptBtn.onmouseover = () => acceptBtn.style.background = '#2563eb';
-              acceptBtn.onmouseout = () => acceptBtn.style.background = '#3b82f6';
-              acceptBtn.onclick = () => {
-                const currentEditor = globalEditorRef.current;
-                if (currentEditor) {
-                  currentEditor.commands.unsetAIHighlight();
-                }
-                (controls as any).cleanup?.() || controls.remove();
-              };
-
-              controls.appendChild(undoBtn);
-              controls.appendChild(acceptBtn);
-              document.body.appendChild(controls);
-
-              // Remove controls on outside click
-              const handleOutsideClick = (e: MouseEvent) => {
-                const target = e.target as Node;
-                
-                // Check if clicked on controls
-                if (controls.contains(target)) {
-                  return;
-                }
-                
-                // Check if clicked on highlighted text
-                const currentHighlight = document.querySelector('[data-ai-highlight="true"]');
-                if (currentHighlight && currentHighlight.contains(target)) {
-                  return;
-                }
-                
-                // Clicked outside - accept and cleanup
-                const currentEditor = globalEditorRef.current;
-                if (currentEditor) {
-                  // Just remove the highlight, don't need to select
-                  currentEditor.commands.unsetAIHighlight();
-                }
-                
-                // Use cleanup function if available
-                (controls as any).cleanup?.() || controls.remove();
-              };
-              
-              // Use mousedown for more reliable detection, add on next frame
-              requestAnimationFrame(() => {
-                document.addEventListener('mousedown', handleOutsideClick);
-              });
-
-              // Update position on scroll
-              const updatePosition = () => {
-                const newRect = highlighted.getBoundingClientRect();
-                controls.style.top = `${newRect.top - 36}px`;
-                controls.style.left = `${newRect.left}px`;
-              };
-              window.addEventListener('scroll', updatePosition, true);
-
-              // Cleanup function to remove controls and listeners
-              const cleanup = () => {
-                controls.remove();
-                document.removeEventListener('mousedown', handleOutsideClick);
-                window.removeEventListener('scroll', updatePosition, true);
-              };
-              
-              // Store cleanup on the controls element for access by buttons
-              (controls as any).cleanup = cleanup;
-              
-              // Auto-remove after 20 seconds
-              setTimeout(() => {
-                cleanup();
-                const currentEditor = globalEditorRef.current;
-                if (currentEditor) {
-                  currentEditor.commands.unsetAIHighlight();
-                }
-              }, 20000);
-            };
-
-            // Show controls after a short delay to ensure highlight is rendered
-            setTimeout(createControls, 100);
-
-            return { 
-              success: true, 
-              message: `Successfully replaced content at positions ${startOffset}-${endOffset} with new HTML content`
-            };
-          } catch (error) {
-            console.error("Failed to edit enhanced note:", error);
-            return { success: false, error: `Failed to update the enhanced note: ${error instanceof Error ? error.message : 'Unknown error'}` };
-          }
-        },
+      // Create tools using the refactored tool factories
+      const searchTool = createSearchSessionTool(userId);
+      const editEnhancedNoteTool = createEditEnhancedNoteTool({
+        sessionId,
+        sessions,
+        selectionData,
       });
 
       const abortController = new AbortController();
